Added ITempDataProvider and CookieTempDataProvider
